remove CUDA ext
