+++ src/cuda/cuda_ptx_task.jl
@@ -1,19 +1,12 @@
-function ndarray_cuda_type(A::NDArray{T,N}) where {T,N}
-    if N == 1
-        CuDeviceVector{T,1}
-    elseif N == 2
-        CuDeviceMatrix{T,1}
-    else
-        CuDeviceArray{T,N,1}
-    end
+export @cuda_task, @launch, CUDATask
+
+struct CUDATask
+    func::String
+    argtypes::NTuple{N,Type} where {N}
 end
 
-function ndarray_cuda_type(arg::T) where {T}
-    Base.isbits(arg) || throw(ArgumentError("Unsupported argument type: $(typeof(arg))"))
-    typeof(arg)
-end
-
-cuNumeric.map_ndarray_cuda_types(args...) = tuple(map(ndarray_cuda_type, args)...)
+#! JUST PASS TYPES HERE INSTEAD OF CALLING typeof()
+map_ndarray_cuda_types(args...) = tuple(map(ndarray_cuda_type, typeof.(args))...)
 
 function to_stdvec(::Type{T}, vec) where {T}
     stdvec = CxxWrap.StdVector{T}()
@@ -53,12 +46,12 @@ function add_padding(arr::NDArray, i::Int64; copy=false)
 end
 
 function check_sz!(arr, maxshape; copy=false)
-    sz = cuNumeric.size(arr)
+    sz = size(arr)
     if maxshape != nothing
         # currently require all ndarray inputs to be equal
         alligned_equal_size = sz == maxshape
         if !alligned_equal_size
-            cuNumeric.add_padding(arr, maxshape; copy=copy)
+            add_padding(arr, maxshape; copy=copy)
             new_size = padded_shape(arr)
             @warn "[Padding Added] $sz output is now $new_size"
         end
@@ -79,7 +72,7 @@ function nda_to_logical_array(arr::NDArray{T,N}) where {T,N}
     return Legate.LogicalArray{T,N}(st_handle, size(arr))
 end
 
-function Launch(kernel::cuNumeric.CUDATask, inputs::Tuple{Vararg{NDArray}},
+function Launch(kernel::CUDATask, inputs::Tuple{Vararg{NDArray}},
     outputs::Tuple{Vararg{NDArray}}, scalars::Tuple{Vararg{Any}}; blocks, threads)
 
     # we find the largest input/output.
@@ -125,7 +118,7 @@ function Launch(kernel::cuNumeric.CUDATask, inputs::Tuple{Vararg{NDArray}},
     Legate.submit_auto_task(rt, task)
 end
 
-function cuNumeric.launch(kernel::cuNumeric.CUDATask, inputs, outputs, scalars; blocks, threads)
+function launch(kernel::CUDATask, inputs, outputs, scalars; blocks, threads)
     Launch(kernel,
         isa(inputs, Tuple) ? inputs : (inputs,),
         isa(outputs, Tuple) ? outputs : (outputs,),
@@ -135,7 +128,7 @@ function cuNumeric.launch(kernel::cuNumeric.CUDATask, inputs, outputs, scalars; 
     )
 end
 
-function cuNumeric.ptx_task(ptx::String, kernel_name)
+function ptx_task(ptx::String, kernel_name)
     rt = Legate.get_runtime()
     lib = cuNumeric.get_lib() # grab lib of legate app
     # this taskid is directly tied to cpp code in our setup
@@ -147,6 +140,42 @@ function cuNumeric.ptx_task(ptx::String, kernel_name)
     Legate.submit_auto_task(rt, task)
 end
 
+"""
+    @cuda_task(f(args...))
+
+Compile a Julia GPU kernel to PTX, register it with the Legate runtime,
+and return a `CUDATask` object for later launch.
+
+# Arguments
+- `f` — The name of the Julia CUDA.jl GPU kernel function to compile.
+- `args...` — Example arguments to the kernel, used to determine the
+  argument type signature when generating PTX.
+
+# Description
+This macro automates the process of:
+1. Inferring the CUDA argument types for the given `args` using
+   `map_ndarray_cuda_types`.
+2. Using `CUDA.code_ptx` to compile the specified GPU kernel
+   (`f`) into raw PTX text for the inferred types.
+3. Extracting the kernel's function symbol name from the PTX using
+   `extract_kernel_name`.
+4. Registering the compiled PTX and kernel name with the Legate runtime
+   via `ptx_task`, making it available for GPU execution.
+5. Returning a `CUDATask` struct that stores the kernel name and type signature,
+   which can be used to configure and launch the kernel later.
+
+# Notes
+- The `args...` are not executed; they are used solely for type inference.
+- This macro is intended for use with the Legate runtime and
+  assumes a CUDA context is available.
+- Make sure your kernel code is GPU-compatible and does not rely on
+  unsupported Julia features.
+
+# Example
+```julia
+mytask = @cuda_task my_kernel(A, B, C)
+```
+"""
 macro cuda_task(call_expr)
     cuNumeric.assert_experimental()
 
@@ -155,21 +184,58 @@ macro cuda_task(call_expr)
 
     esc(quote
         local _buf = IOBuffer()
-        local _types = cuNumeric.map_ndarray_cuda_types($(fargs...))
+        local _types = map_ndarray_cuda_types($(fargs...))
         # generate ptx using CUDA.jl
-        CUDA.code_ptx(_buf, $fname, _types; raw=false, kernel=true)
+        CUDATools.code_ptx(_buf, $fname, _types; raw=false, kernel=true)
 
         local _ptx = String(take!(_buf))
-        local _func_name = cuNumeric.extract_kernel_name(_ptx)
+        local _func_name = extract_kernel_name(_ptx)
 
         # issue ptx_task within legate runtime to register cufunction ptr with cucontext
-        cuNumeric.ptx_task(_ptx, _func_name)
+        ptx_task(_ptx, _func_name)
 
         # create a cuNumeric.CUDAtask that stores some info for a launch config
-        cuNumeric.CUDATask(_func_name, _types)
+        CUDATask(_func_name, _types)
     end)
 end
 
+"""
+    @launch(; task, blocks=(1,), threads=(256,), inputs=(), outputs=(), scalars=())
+
+Launch a GPU kernel (previously registered via [`@cuda_task`](@ref))  through the Legate runtime.
+
+# Keywords
+- `task` — A `CUDATask` object, typically returned by [`@cuda_task`](@ref).
+- `blocks`  — Tuple or single element specifying the CUDA grid dimensions. Defaults to `(1,)`.
+- `threads` — Tuple or single element specifying the CUDA block dimensions. Defaults to `(256,)`.
+- `inputs`  — Tuple or single element of input NDArray objects.
+- `outputs` — Tuple or single element of output NDArray objects.
+- `scalars` — Tuple or single element of scalar values.
+
+# Description
+The `@launch` macro validates the provided keywords, ensuring only
+the allowed set (`:task`, `:blocks`, `:threads`, `:inputs`, `:outputs`, `:scalars`)
+are present. It then expands to a call to `cuNumeric.launch`,
+passing the given arguments to the Legate runtime for execution.
+
+This macro is meant to provide a concise, declarative syntax for
+launching GPU kernels, separating kernel compilation (via `@cuda_task`)
+from execution configuration.
+
+# Notes
+- `task` **must** be a kernel registered with the runtime, usually from `@cuda_task`.
+- All keyword arguments must be specified as assignments, e.g. `blocks=(2,2)` not positional arguments.
+- Defaults are chosen for single-block, 256-thread 1D launches.
+- The macro escapes its body so that the values of inputs/outputs/scalars are captured
+  from the surrounding scope at macro expansion time.
+
+# Example
+```julia
+mytask = @cuda_task my_kernel(A, B, C)
+
+@launch task=mytask blocks=(8,8) threads=(32,32) inputs=(A, B) outputs=(C)
+```
+"""
 macro launch(args...)
     cuNumeric.assert_experimental()
